Added partial createTextures build task [WIP].
* This requires further work as previously the task generated a class
in order to define it with an array of image references. In the course of
simplifying the build the list should be changed to a property file.
The generated class is now statically defined and it needs to populate
itself using the generated list in order to regain it's functionality.
* Also added the previous ant build files as commented code in the gradle
build file as placeholders for missing build tasks.